+++ src/main/java/com/emazon/ms_stock/application/dto/input/ArticleReqDTO.java
@@ -1,4 +1,4 @@
-package com.emazon.ms_stock.application.dto;
+package com.emazon.ms_stock.application.dto.input;
 
 import jakarta.validation.constraints.*;
 import lombok.*;
